Rename jwt as jwt-check
