+++ src/route/genres.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import jwtAuthz from 'express-jwt-authz';
 import * as Yup from 'yup';
 import { Genre } from '../entity/genre';
-import jwt from '../middleware/jwt';
+import jwtCheck from '../middleware/jwt-check';
 import HttpError from '../util/http-error';
 
 const router = express.Router();
@@ -17,7 +17,7 @@ router.get('/', async (req, res) => {
   res.send({ data: genres });
 });
 
-router.post('/', jwt, jwtAuthz(['write:genres']), async (req, res) => {
+router.post('/', jwtCheck, jwtAuthz(['write:genres']), async (req, res) => {
   const value = await schema.validate(req.body, { stripUnknown: true });
 
   const genre = Genre.create({ ...value });
@@ -26,7 +26,7 @@ router.post('/', jwt, jwtAuthz(['write:genres']), async (req, res) => {
   res.send({ data: genre });
 });
 
-router.put('/:id', jwt, jwtAuthz(['update:genres']), async (req, res) => {
+router.put('/:id', jwtCheck, jwtAuthz(['update:genres']), async (req, res) => {
   const value = await schema.validate(req.body, { stripUnknown: true });
 
   const result = await Genre.update(parseInt(req.params.id, 10), { ...value });
@@ -37,14 +37,19 @@ router.put('/:id', jwt, jwtAuthz(['update:genres']), async (req, res) => {
   res.send({ rows_affected: result.affected });
 });
 
-router.delete('/:id', jwt, jwtAuthz(['delete:genres']), async (req, res) => {
-  const result = await Genre.delete(parseInt(req.params.id, 10));
-  if (!result.affected) {
-    throw new HttpError(404, 'Genre not found');
-  }
+router.delete(
+  '/:id',
+  jwtCheck,
+  jwtAuthz(['delete:genres']),
+  async (req, res) => {
+    const result = await Genre.delete(parseInt(req.params.id, 10));
+    if (!result.affected) {
+      throw new HttpError(404, 'Genre not found');
+    }
 
-  res.send({ rows_affected: result.affected });
-});
+    res.send({ rows_affected: result.affected });
+  }
+);
 
 router.get('/:id', async (req, res) => {
   const genre = await Genre.findOne(parseInt(req.params.id, 10));
