+++ src/route/movies.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import jwtAuthz from 'express-jwt-authz';
 import * as Yup from 'yup';
 import { Movie } from '../entity/movie';
-import jwt from '../middleware/jwt';
+import jwtCheck from '../middleware/jwt-check';
 import HttpError from '../util/http-error';
 import knex from '../util/knex';
 
@@ -21,7 +21,7 @@ router.get('/', async (req, res) => {
   res.send({ data: movies });
 });
 
-router.post('/', jwt, jwtAuthz(['write:movies']), async (req, res) => {
+router.post('/', jwtCheck, jwtAuthz(['write:movies']), async (req, res) => {
   const value = await schema.validate(req.body, { stripUnknown: true });
 
   const movie = Movie.create(value);
@@ -30,7 +30,7 @@ router.post('/', jwt, jwtAuthz(['write:movies']), async (req, res) => {
   res.send({ data: movie });
 });
 
-router.put('/:id', jwt, jwtAuthz(['update:movies']), async (req, res) => {
+router.put('/:id', jwtCheck, jwtAuthz(['update:movies']), async (req, res) => {
   const value = await schema.validate(req.body, { stripUnknown: true });
 
   const genre = await knex('genres').where('id', value.genre_id).first();
@@ -48,14 +48,19 @@ router.put('/:id', jwt, jwtAuthz(['update:movies']), async (req, res) => {
   res.send({ data: result });
 });
 
-router.delete('/:id', jwt, jwtAuthz(['delete:movies']), async (req, res) => {
-  const result = await Movie.delete(parseInt(req.params.id, 10));
-  if (!result.affected) {
-    throw new HttpError(404, 'Movie not found');
-  }
+router.delete(
+  '/:id',
+  jwtCheck,
+  jwtAuthz(['delete:movies']),
+  async (req, res) => {
+    const result = await Movie.delete(parseInt(req.params.id, 10));
+    if (!result.affected) {
+      throw new HttpError(404, 'Movie not found');
+    }
 
-  res.send({ rows_affected: result.affected });
-});
+    res.send({ rows_affected: result.affected });
+  }
+);
 
 router.get('/:id', async (req, res) => {
   const movie = await Movie.findOne(parseInt(req.params.id, 10));
